moved UseCase classes to own package
