+++ app/src/main/java/com/peternaggschga/gwent/domain/cases/RowStateUseCase.java
@@ -1,4 +1,4 @@
-package com.peternaggschga.gwent.domain;
+package com.peternaggschga.gwent.domain.cases;
 
 import androidx.annotation.NonNull;
