+++ app/src/main/java/com/peternaggschga/gwent/ui/GameBoardViewModel.java
@@ -11,8 +11,8 @@ import androidx.lifecycle.viewmodel.ViewModelInitializer;
 import com.peternaggschga.gwent.GwentApplication;
 import com.peternaggschga.gwent.data.RowType;
 import com.peternaggschga.gwent.data.UnitRepository;
-import com.peternaggschga.gwent.domain.BurnUnitsUseCase;
-import com.peternaggschga.gwent.domain.RowStateUseCase;
+import com.peternaggschga.gwent.domain.cases.BurnUnitsUseCase;
+import com.peternaggschga.gwent.domain.cases.RowStateUseCase;
 
 import java.util.HashMap;
 import java.util.Map;
